Fix: Ansible Linux (#26)
# What I did
- i fixed it so that ansible is run only once and on your machine
- will not work on Windows, only Linux/macOS

I've also identified the problem with Steps 18 and 19. It's actually
related to the Flannel implementation. The main issue seems to be
timing. Our patch command runs before the Flannel DaemonSet exists, so
it fails silently and Flannel uses the wrong network interface. We
should wait for it to exist or something. I don't have time to fix it
and frankly, it's not my assigned task.

# How to run
`vagrant up` should show ctrl being provisioned properly. The nodes will
fail because of Step 15.