Feat/ass4 3 (#55)
This PR completes: 
Additional Use Case: Istio is a rich tool and its features allow other
uses that go beyond simple request routing.
Implement any of the following ideas (or one of your own, as long as it
is comparable in complexity).
Shadow launch: A second instance of the model-service is deployed that
uses a newer version of the model (can
be simple re-release without changes). The deployment will mirror
existing traffic and submit it to both instances.

Check the "With Istio Shadow Launch (model mirroring)" part of the
README.

---------

Co-authored-by: Nicholas <nicholaswu2003@gmail.com>